Uses file extension to clarify the install script isn't a directory (#16)
It is normal in UNIX to not add a shell script suffix even if it is
common in projects that have both windows and UNIX commands in them.

For example, the canonical name for the Docker HEALTHCHECK command is
`docker-healthcheck`, not `docker-healthcheck.sh`.

There's a special case @jcchavezs noticed when looking at how the
installation layer is setup. The word `install` is both used for a
root directory `/install` and also a script that populates that,
`/tmp/install`. This distracted him from other work and that's an
anti-goal.

Instead of deciding a different name for `/install` or a different
name for the `/tmp/install` script, this appends an otherwise
unnecessary `.sh` to the install script to avoid the problem without
creating a new naming problem.